refactor: create module for stock methods
